Handling a public key storage
