+++ protocol/auth/src/main/kotlin/com/walletconnect/auth/client/mapper/ClientMapper.kt
@@ -8,7 +8,7 @@ import com.walletconnect.android.internal.common.model.SDKError
 import com.walletconnect.android.internal.common.model.Validation
 import com.walletconnect.android.internal.common.signing.cacao.Cacao
 import com.walletconnect.android.internal.common.signing.cacao.CacaoType
-import com.walletconnect.android.verify.data.model.VerifyContext
+import com.walletconnect.android.verify.model.VerifyContext
 import com.walletconnect.auth.client.Auth
 import com.walletconnect.auth.common.model.AuthResponse
 import com.walletconnect.auth.common.model.Events
