+++ protocol/auth/src/main/kotlin/com/walletconnect/auth/common/model/Events.kt
@@ -3,7 +3,7 @@
 package com.walletconnect.auth.common.model
 
 import com.walletconnect.android.internal.common.model.type.EngineEvent
-import com.walletconnect.android.verify.data.model.VerifyContext
+import com.walletconnect.android.verify.model.VerifyContext
 
 internal sealed class Events : EngineEvent {
     data class OnAuthRequest(val id: Long, val pairingTopic: String, val payloadParams: PayloadParams, val verifyContext: VerifyContext) : Events()
