+++ protocol/auth/src/main/kotlin/com/walletconnect/auth/use_case/calls/GetListOfVerifyContextsUseCase.kt
@@ -1,7 +1,7 @@
 package com.walletconnect.auth.use_case.calls
 
 import com.walletconnect.android.internal.common.storage.verify.VerifyContextStorageRepository
-import com.walletconnect.android.verify.data.model.VerifyContext
+import com.walletconnect.android.verify.model.VerifyContext
 import kotlinx.coroutines.supervisorScope
 
 internal class GetListOfVerifyContextsUseCase(private val verifyContextStorageRepository: VerifyContextStorageRepository) : GetListOfVerifyContextsUseCaseInterface {
